refactor(game): reorganize packages following DDD architecture
