+++ src/main/java/com/itacademy/blackjack/game/domain/model/GameResult.java
@@ -1,4 +1,4 @@
-package com.itacademy.blackjack.game.model;
+package com.itacademy.blackjack.game.domain.model;
 
 public enum GameResult {
     PLAYER_WINS, PUSH, CRUPIER_WINS, NO_RESULTS_YET, BLACKJACK
