+++ src/main/java/com/itacademy/blackjack/game/domain/model/GameStatus.java
@@ -1,4 +1,4 @@
-package com.itacademy.blackjack.game.model;
+package com.itacademy.blackjack.game.domain.model;
 
 public enum GameStatus {
     CREATED,STARTED,PLAYER_TURN,CRUPIER_TURN,FINISHED
