+++ src/main/java/com/itacademy/blackjack/game/domain/model/PlayerStatus.java
@@ -1,4 +1,4 @@
-package com.itacademy.blackjack.game.model;
+package com.itacademy.blackjack.game.domain.model;
 
 public enum PlayerStatus {
     ACTIVE,
